requirements.txt
nodriver>=0.40
feedgen==1.0.0
python-dotenv==1.1.0
google-genai>=1.0.0
yt-dlp>=2024.0
